add spotify.js with query method
