+++ app/server.js
@@ -1,15 +1,18 @@
 require("dotenv").config();
 
 const express = require("express");
-const returnAPIKey = require("./key");
+const spotify = require("./spotify");
 const path = require("path");
 
 async function run() {
   // spotify bearer token
   var client_id = process.env.CLIENT_ID;
   var client_secret = process.env.CLIENT_SECRET;
-  var bearer = await returnAPIKey(client_id, client_secret);
+  var bearer = await spotify.returnBearer(client_id, client_secret);
   console.log(bearer);
+
+  var searchResults = await spotify.search(bearer, "miles%20davis&type=artist");
+  console.log(searchResults);
 
   // server
   const PORT = 8888;
